+++ Core/Inc/dribbler_board/peripherals/dribbler.h
@@ -37,7 +37,7 @@ void dribbler_Init();
 // Initializes the motor driver for the PWM
 void dribbler_motor_Init();
 // The current value we set, limits acceleration
-void dribbler_setCurrentLimit(uint16_t value);
+void dribbler_setCurrentLimit(float value);
 // Larger current implies we have the ball
 uint32_t dribbler_getCurrent();
 // If the current is above a certain threshold, it implies we have the ball
